Fix : ShareModal 컴포넌트로 분리:
